Renamed to score_opt_partition.hpp
