+++ include/wand_data_compressed.hpp
@@ -5,7 +5,7 @@
 #include "binary_freq_collection.hpp"
 #include "bm25.hpp"
 #include "util.hpp"
-#include "score_partitioning.hpp"
+#include "score_opt_partition.hpp"
 #include "bitvector_collection.hpp"
 #include "configuration.hpp"
